+++ src/MemUsage.cxx
@@ -102,7 +102,7 @@ MemUsage::event MemUsage::operator[](int ind) const
 std::string MemUsage::summary() const
 {
     stringstream ss;
-    for (int ind=0; ind<m_events.size(); ++ind) {
+    for (size_t ind=0; ind<m_events.size(); ++ind) {
 	ss << this->emit(ind) << "\n";
     }
     return ss.str();
