+++ test/test_boost_multi_array.cxx
@@ -11,8 +11,8 @@ int main()
     typedef boost::multi_array<double, 2> array_type;
     typedef array_type::index index;
 
-    int l_size = 10;
-    int t_size = 3;
+    size_t l_size = 10;
+    size_t t_size = 3;
     array_type ar(boost::extents[l_size][t_size]);
     cerr << "Dimensions: " << ar.dimensionality << endl;
     cerr << "Dimension 0 is size " << ar.shape()[0] << endl;
@@ -20,13 +20,13 @@ int main()
 
     Assert(l_size == ar.shape()[0]);
     Assert(t_size == ar.shape()[1]);
-    for (int l_ind=0; l_ind < l_size; ++l_ind) {
-	for (int t_ind=0; t_ind < t_size; ++t_ind) {
+    for (size_t l_ind=0; l_ind < l_size; ++l_ind) {
+	for (size_t t_ind=0; t_ind < t_size; ++t_ind) {
 	    ar[l_ind][t_ind] = l_ind*t_ind;
 	}
     }
-    for (int t_ind=0; t_ind < t_size; ++t_ind) {
-	for (int l_ind=0; l_ind < l_size; ++l_ind) {
+    for (size_t t_ind=0; t_ind < t_size; ++t_ind) {
+	for (size_t l_ind=0; l_ind < l_size; ++l_ind) {
 	    cerr << "\t[" << l_ind<< "][" << t_ind << "] = " << ar[l_ind][t_ind];
 	}
 	cerr << endl;
